Backend/requirements.txt
flask
python-dotenv
psycopg2-binary
flask_sqlalchemy
flask_migrate
httpagentparser
flask-cors